fix: change code folder structure
