+++ src/data-access/config/dataSource.ts
@@ -1,6 +1,6 @@
 import 'reflect-metadata'
 import { DataSource } from 'typeorm'
-import { User } from './entity/user.entity'
+import { User } from '../entity/user.entity'
 
 export const dataSource = new DataSource({
     type: 'postgres',
