+++ src/data-access/repositories/user.repository.ts
@@ -2,7 +2,7 @@ import {IUserRepository} from "./user.repository.interface";
 import {Repository} from "typeorm/repository/Repository";
 import {User} from "../entity/user.entity";
 import {v4 as uuidv4} from "uuid";
-import {dataSource} from "../dataSource";
+import {dataSource} from "../config/dataSource";
 
 export class UserRepository implements IUserRepository {
     constructor(private repository: Repository<User>) {}
