+++ src/router/controllers/rate.controller.ts
@@ -1,8 +1,8 @@
 import {Request, Response} from 'express'
-import logger from '../helpers/logger'
-import {IRateService} from "../services/rate/rate.service.interface";
-import rateService from "../services/rate/rate.service";
-import {errorHandler} from "../error/handler/error.handler";
+import logger from '../../helpers/logger'
+import {IRateService} from "../../service/services/rate/rate.service.interface";
+import rateService from "../../service/services/rate/rate.service";
+import {errorHandler} from "../../error/handler/error.handler";
 
 export class RateController {
