+++ src/service/schedule/interface/sendEmail.interface.ts
