+++ src/service/services/rate/rate.service.interface.ts
