+++ src/service/services/user/user.service.interface.ts
@@ -1,0 +1,6 @@
+import { User } from '../../../data-access/entity/user.entity';
+
+export interface IUserService {
+    subscribeEmail(email: string): Promise<User>;
+    getAllUsers(): Promise<User[]>;
+}
