--- src/services/user/user.service.interface.ts
@@ -1,6 +1,0 @@
-import { User } from '../../entity/user.entity';
-
-export interface IUserService {
-    subscribeEmail(email: string): Promise<void>;
-    getAllUsers(): Promise<User[]>;
-}
