+++ src/test/service/user.service.test.ts
@@ -1,15 +1,15 @@
 import sinon from 'sinon';
-import userService from '../../services/user/user.service'
-import { User } from '../../entity/user.entity'
-import { UserAlreadyExistsError } from '../../error/user.error';
-import {dataSource} from "../../dataSource";
+import userService from '../../service/services/user/user.service'
+import { User } from '../../data-access/entity/user.entity'
 import {FIRST_EMAIL, FIRST_ID, SECOND_EMAIL, SECOND_ID} from "../mock/user.const";
+import userRepository, {UserRepository} from "../../data-access/repositories/user.repository";
+import UserAlreadyExistError from "../../error/types/userAlreadyExist.error";
 
 describe('UserService', () => {
-    let userRepositoryStub: sinon.SinonStubbedInstance<any>;
+    let userRepositoryStub: sinon.SinonStubbedInstance<UserRepository>;
 
     beforeEach(() => {
-        userRepositoryStub = sinon.stub(dataSource.getRepository(User));
+        userRepositoryStub = sinon.stub(userRepository);
     });
 
     afterEach(() => {
@@ -18,26 +18,21 @@ describe('UserService', () => {
 
     describe('test subscribeEmail method', () => {
         it('should create a new user if user with given email does not exist', async () => {
-            userRepositoryStub.findOne.resolves(null);
-            userRepositoryStub.create.returns({ id: FIRST_ID, email: FIRST_EMAIL });
-            userRepositoryStub.save.resolves({ id: FIRST_ID, email: FIRST_EMAIL });
+            userRepositoryStub.saveByEmail.resolves({ id: FIRST_ID, email: FIRST_EMAIL });
 
-            const user = await userService.subscribeEmail(FIRST_EMAIL);
+            const user: User = await userService.subscribeEmail(FIRST_EMAIL);
 
-            expect(user.id).toBe(FIRST_ID);
             expect(user.email).toBe(FIRST_EMAIL);
         });
 
         it('should throw an error if email already exists', async () => {
-            userRepositoryStub.findOne.resolves({ id: FIRST_ID, email: FIRST_EMAIL });
+            userRepositoryStub.getByEmail.resolves({ id: FIRST_ID, email: FIRST_EMAIL });
 
-            await expect(userService.subscribeEmail(FIRST_EMAIL)).rejects.toThrow(UserAlreadyExistsError);
+            await expect(userService.subscribeEmail(FIRST_EMAIL)).rejects.toThrow(UserAlreadyExistError);
         });
 
         it('should throw an error if saving user fails', async () => {
-            userRepositoryStub.findOne.resolves(null);
-            userRepositoryStub.create.returns({ id: FIRST_ID, email: FIRST_EMAIL });
-            userRepositoryStub.save.rejects(new Error('Database error'));
+            userRepositoryStub.saveByEmail.rejects(new Error('Database error'));
 
             await expect(userService.subscribeEmail(FIRST_EMAIL)).rejects.toThrow('Error creating user: Database error');
         });
@@ -49,7 +44,7 @@ describe('UserService', () => {
                 { id: FIRST_ID, email: FIRST_EMAIL },
                 { id: SECOND_ID, email: SECOND_EMAIL }
             ];
-            userRepositoryStub.find.resolves(users);
+            userRepositoryStub.getAll.resolves(users);
 
             const result = await userService.getAllUsers();
 
@@ -57,7 +52,7 @@ describe('UserService', () => {
         });
 
         it('should throw an error if fetching users fails', async () => {
-            userRepositoryStub.find.rejects(new Error('Database error'));
+            userRepositoryStub.getAll.rejects(new Error('Database error'));
 
             await expect(userService.getAllUsers()).rejects.toThrow('Error getting all users email: Database error');
         });
